requirements.txt
click==7.1.2
decorator==4.4.2
dnspython==1.16.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
nanoid==2.0.0
pymongo==3.10.1
python-dotenv==0.14.0
six==1.15.0
validators==0.15.0
Werkzeug==1.0.1
